chore: move find and focus custom modules to tabulator dir
