+++ log-viewer/src/tabulator/module/Find.ts
