+++ log-viewer/src/tabulator/module/MiddleRowFocus.ts
@@ -2,7 +2,7 @@
  * Copyright (c) 2024 Certinia Inc. All rights reserved.
  */
 import { Module, type RowComponent, type Tabulator } from 'tabulator-tables';
-import type { LogEvent } from '../../../core/log-parser/LogEvents.js';
+import type { LogEvent } from '../../core/log-parser/LogEvents.js';
 
 type TimedNodeProp = { originalData: LogEvent };
